+++ packages/expo-dev-launcher/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Add expo-modules and ReactDelegate-based automatic setup on iOS. ([#16190](https://github.com/expo/expo/pull/16190) by [@esamelson](https://github.com/esamelson))
 - Add support for auto-setup with updates integration on iOS. ([#16230](https://github.com/expo/expo/pull/16230) by [@esamelson](https://github.com/esamelson))
+- Add support for th React Native `0.67.X`.
 
 ### 🐛 Bug fixes